Move  impl and add more a generic mnemonic creation method
